Move PegoutValue into new peg module
